Fix bug where QEMU fast-forwards when it shouldn't
Basically, the main loop doesn't realize that the VM has stopped when
it asks for warping to be computed, and that means it jumps far ahead
in time when it shouldn't.

In the worst case, the next timer isn't until INT64_MAX, and QEMU ends
up hanging, because it skips too far into the future.